[refector] 도서, 서재 관련 레이어 분리, 서재 Dialog 데이터 흐름 수정, ui 레이어 패키지 구분

+++ BookChat/app/src/main/java/com/example/bookchat/data/response/BookShelfMeta.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class BookShelfMeta(
     @SerializedName("totalElements")
-    val totalElements: Long,
+    val totalElements: Int,
     @SerializedName("totalPages")
     val totalPages: Int,
     @SerializedName("pageSize")
